requirements.txt
sunpy[all]
wandb
pytorch-lightning==1.9.3
numpy==1.26.2
itipy